requirements.txt
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Login==0.4.1
flask-mongoengine==0.9.5
Flask-WTF==0.14.2
google-api-python-client==1.7.11
google-auth==1.7.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mongoengine==0.18.2
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pymongo==3.9.0
pyTelegramBotAPI==3.6.6
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
schedule==0.6.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
